# BytesToString: add benchmarks against string(bytes)