After running
```shell
$ go test -run=^$ -bench=. -count=10
```

and separating the content into:
* lib.txt for the benchmark output for BytesToString
* ordinary.txt for the usual string(bytes) conversion

the benchmark shows the improvement of switching to BytesToString:

```shell
$ benchstat ordinary.txt lib.txt
name             old time/op    new time/op    delta
BytesToString-8     104ns ± 5%      54ns ± 2%  -47.60%  (p=0.000 n=10+9)

name             old alloc/op   new alloc/op   delta
BytesToString-8      128B ± 0%       32B ± 0%  -75.00%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
BytesToString-8      4.00 ± 0%      2.00 ± 0%  -50.00%  (p=0.000 n=10+10)
```